matched csnx storybook webpack config in preparation for migration
